+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change Log
 
-## Unreleased 0.0.3
+## [0.0.3](https://github.com/shoorick/codeception-test-adapter/releases/tag/v0.0.3) - 2025-12-21
 
+- Add `codeceptionTestAdapter.*` settings
 - Decrease refresh delay
 - Support multiple workspace folders